+++ lib/src/core/foundation/common/render_object.dart
@@ -1,8 +1,8 @@
 import 'dart:html';
 
 import 'package:meta/meta.dart';
-import 'package:rad/src/core/services/debug.dart';
-import 'package:rad/src/core/services/framework_registry.dart';
+import 'package:rad/src/core/utilities/debug.dart';
+import 'package:rad/src/core/utilities/services_registry.dart';
 import 'package:rad/src/core/enums.dart';
 import 'package:rad/src/core/foundation/common/build_context.dart';
 import 'package:rad/src/core/foundation/scheduler/scheduler.dart';
@@ -21,7 +21,7 @@ abstract class RenderObject {
   /// Get task scheduler for app instance that's enclosing the current context.
   ///
   Scheduler get scheduler =>
-      FrameworkRegistry.instance.getTaskScheduler(context);
+      ServicesRegistry.instance.getTaskScheduler(context);
 
   /*
   |--------------------------------------------------------------------------
