+++ lib/src/core/foundation/services.dart
@@ -1,0 +1,9 @@
+import 'package:rad/src/core/foundation/scheduler/scheduler.dart';
+import 'package:rad/src/core/foundation/walker/walker.dart';
+
+/// Services available.
+///
+class Services {
+  final walker = Walker();
+  final scheduler = Scheduler();
+}
