+++ lib/src/core/functions.dart
@@ -1,16 +1,15 @@
 import 'dart:html';
 
-import 'package:rad/src/core/services/debug.dart';
+import 'package:rad/rad.dart';
+import 'package:rad/src/core/enums.dart';
+import 'package:rad/src/core/foundation/services.dart';
+import 'package:rad/src/core/utilities/debug.dart';
 import 'package:rad/src/core/foundation/framework.dart';
-import 'package:rad/src/core/services/framework_registry.dart';
-import 'package:rad/src/core/constants.dart';
-import 'package:rad/src/core/foundation/common/build_context.dart';
-import 'package:rad/src/core/foundation/common/debug_options.dart';
+import 'package:rad/src/core/utilities/services_registry.dart';
 import 'package:rad/src/core/foundation/scheduler/tasks/widgets_build_task.dart';
-import 'package:rad/src/widgets/abstract/widget.dart';
 import 'package:rad/src/widgets/utils/common_props.dart';
 
-void startApp({
+void fnStartApp({
   required Widget app,
   required String targetSelector,
   VoidCallback? beforeMount,
@@ -18,7 +17,41 @@ void startApp({
 }) {
   /*
   |--------------------------------------------------------------------------
-  | Pre-checks
+  | Update debug options
+  |--------------------------------------------------------------------------
+  */
+
+  Debug.update(debugOptions);
+
+  /*
+  |--------------------------------------------------------------------------
+  | Create root context
+  |--------------------------------------------------------------------------
+  */
+
+  var rootContext = BuildContext.bigBang(targetSelector);
+
+  /*
+  |--------------------------------------------------------------------------
+  | Register services
+  |--------------------------------------------------------------------------
+  */
+
+  var services = Services();
+
+  ServicesRegistry.instance.registerServices(rootContext, services);
+
+  /*
+  |--------------------------------------------------------------------------
+  | Spin framework instance for app
+  |--------------------------------------------------------------------------
+  */
+
+  Framework(rootContext, services);
+
+  /*
+  |--------------------------------------------------------------------------
+  | Pre-checks before mount
   |--------------------------------------------------------------------------
   */
 
@@ -30,44 +63,30 @@ void startApp({
 
   /*
   |--------------------------------------------------------------------------
-  | App bootstrap process
+  | Decorate target element
   |--------------------------------------------------------------------------
   */
-
-  // Update debug options.
-
-  Debug.update(debugOptions);
-
-  // Decorate root element.
 
   CommonProps.applyDataAttributes(targetElement, {
     System.attrConcreteType: "Target",
     System.attrRuntimeType: System.contextTypeBigBang,
   });
 
-  // Create framework instance for app.
-
-  var framework = Framework();
-
-  // Create root context for app.
-
-  var rootContext = BuildContext.bigBang(targetSelector);
-
-  // Initialize framework.
-
-  framework.init(rootContext);
-
-  // Fire before mount hook.
+  /*
+  |--------------------------------------------------------------------------
+  | Trigger hooks
+  |--------------------------------------------------------------------------
+  */
 
   beforeMount?.call();
 
-  // Schedule a build task.
+  /*
+  |--------------------------------------------------------------------------
+  | Schedule a build task.
+  |--------------------------------------------------------------------------
+  */
 
-  // Framework is responsible for resgistering a task schedular for provided
-  // root context when its init() method is called. therefore by this point
-  // we assume that our scheduler for context is all set and ready to go.
-
-  var scheduler = FrameworkRegistry.instance.getTaskScheduler(rootContext);
+  var scheduler = ServicesRegistry.instance.getTaskScheduler(rootContext);
 
   scheduler.addTask(
     WidgetsBuildTask(
@@ -75,4 +94,223 @@ void startApp({
       parentContext: rootContext,
     ),
   );
+}
+
+String fnMapDomTag(DomTag tag) {
+  switch (tag) {
+    case DomTag.header:
+      return "header";
+
+    case DomTag.footer:
+      return "footer";
+
+    case DomTag.navigation:
+      return "nav";
+
+    case DomTag.division:
+      return "div";
+
+    case DomTag.span:
+      return "span";
+
+    case DomTag.anchor:
+      return "a";
+
+    case DomTag.blockquote:
+      return "blockquote";
+
+    case DomTag.horizontalRule:
+      return "hr";
+
+    case DomTag.label:
+      return "label";
+
+    case DomTag.iFrame:
+      return "iframe";
+
+    case DomTag.breakLine:
+      return "br";
+
+    case DomTag.image:
+      return "img";
+
+    case DomTag.canvas:
+      return "canvas";
+
+    case DomTag.paragraph:
+      return "p";
+
+    case DomTag.input:
+      return "input";
+
+    case DomTag.form:
+      return "form";
+
+    case DomTag.fieldSet:
+      return "fieldset";
+
+    case DomTag.legend:
+      return "legend";
+
+    case DomTag.idiomatic:
+      return "i";
+
+    case DomTag.strong:
+      return "strong";
+
+    case DomTag.small:
+      return "small";
+
+    case DomTag.subScript:
+      return "sub";
+
+    case DomTag.superScript:
+      return "sup";
+
+    case DomTag.unOrderedList:
+      return "ul";
+
+    case DomTag.listItem:
+      return "li";
+
+    case DomTag.button:
+      return "button";
+
+    case DomTag.select:
+      return "select";
+
+    case DomTag.option:
+      return "option";
+
+    case DomTag.progress:
+      return "progress";
+
+    case DomTag.textArea:
+      return "textarea";
+
+    // headings
+
+    case DomTag.heading1:
+      return "h1";
+
+    case DomTag.heading2:
+      return "h2";
+
+    case DomTag.heading3:
+      return "h3";
+
+    case DomTag.heading4:
+      return "h4";
+
+    case DomTag.heading5:
+      return "h5";
+
+    case DomTag.heading6:
+      return "h6";
+  }
+}
+
+String fnMapDomEventType(DomEventType eventType) {
+  switch (eventType) {
+    case DomEventType.click:
+      return "click";
+
+    case DomEventType.change:
+      return "change";
+
+    case DomEventType.input:
+      return "input";
+
+    case DomEventType.submit:
+      return "submit";
+  }
+}
+
+String fnMapInputType(InputType type) {
+  switch (type) {
+    case InputType.text:
+      return "text";
+
+    case InputType.password:
+      return "password";
+
+    case InputType.file:
+      return "file";
+
+    case InputType.radio:
+      return "radio";
+
+    case InputType.checkbox:
+      return "checkbox";
+
+    case InputType.submit:
+      return "submit";
+  }
+}
+
+String fnMapButtonType(ButtonType buttonType) {
+  switch (buttonType) {
+    case ButtonType.button:
+      return "button";
+
+    case ButtonType.submit:
+      return "submit";
+
+    case ButtonType.reset:
+      return "reset";
+  }
+}
+
+String fnMapFormEncType(FormEncType type) {
+  switch (type) {
+    case FormEncType.applicationXwwwFormUrlEncoded:
+      return "application/x-www-form-urlencoded";
+
+    case FormEncType.multipartFormData:
+      return "multipart/form-data";
+
+    case FormEncType.textPlain:
+      return "text/plain";
+  }
+}
+
+String fnMapFormMethod(FormMethod method) {
+  switch (method) {
+    case FormMethod.post:
+      return "post";
+
+    case FormMethod.get:
+      return "get";
+  }
+}
+
+bool fnIsKeyValueMapEqual(
+  Map<String, String> mapOne,
+  Map<String, String> mapTwo,
+) {
+  if (mapOne.length != mapTwo.length) return false;
+
+  for (var key in mapOne.keys) {
+    if (mapOne[key] != mapTwo[key]) return false;
+  }
+
+  return true;
+}
+
+String fnEncodeKeyValueMap(Map<String, String> valueMap) {
+  var encodedMap = '';
+
+  for (var key in valueMap.keys) {
+    if (key.isNotEmpty) {
+      encodedMap += '/${Uri.encodeComponent(key)}';
+    }
+
+    var value = valueMap[key];
+
+    if (null != value && value.isNotEmpty) {
+      encodedMap += '/${Uri.encodeComponent(valueMap[key]!)}';
+    }
+  }
+
+  return encodedMap;
 }
