+++ lib/src/widgets/html/button.dart
@@ -1,6 +1,7 @@
 import 'dart:html';
 
-import 'package:rad/src/core/services/utils.dart';
+import 'package:rad/src/core/functions.dart';
+import 'package:rad/src/core/utilities/key_generator.dart';
 import 'package:rad/src/core/enums.dart';
 import 'package:rad/src/core/foundation/common/build_context.dart';
 import 'package:rad/src/core/foundation/common/render_object.dart';
@@ -170,7 +171,7 @@ class _ButtonProps {
     }
 
     if (null != props.type) {
-      element.type = Utils.mapButtonType(props.type!);
+      element.type = fnMapButtonType(props.type!);
     }
 
     if (null != props.disabled) {
