+++ lib/src/widgets/navigator.dart
@@ -1,14 +1,15 @@
 import 'dart:html';
 
 import 'package:meta/meta.dart';
-import 'package:rad/src/core/services/debug.dart';
-import 'package:rad/src/core/services/framework_registry.dart';
+import 'package:rad/src/core/functions.dart';
+import 'package:rad/src/core/utilities/debug.dart';
+import 'package:rad/src/core/utilities/services_registry.dart';
 import 'package:rad/src/core/foundation/scheduler/scheduler.dart';
 import 'package:rad/src/core/foundation/scheduler/tasks/widgets_build_task.dart';
 import 'package:rad/src/core/foundation/scheduler/tasks/widgets_manage_task.dart';
 import 'package:rad/src/core/foundation/scheduler/tasks/widgets_update_dependent_task.dart';
 import 'package:rad/src/router/router.dart';
-import 'package:rad/src/core/services/utils.dart';
+import 'package:rad/src/core/utilities/key_generator.dart';
 import 'package:rad/src/core/constants.dart';
 import 'package:rad/src/core/enums.dart';
 import 'package:rad/src/core/foundation/common/build_context.dart';
@@ -294,7 +295,7 @@ class Navigator extends Widget {
     var targetContext = context;
 
     while (true) {
-      var walker = FrameworkRegistry.instance.getTreeWalker(context);
+      var walker = ServicesRegistry.instance.getTreeWalker(context);
 
       var widgetObject = walker.findAncestorWidgetObjectOfType<Navigator>(
         targetContext,
@@ -475,7 +476,7 @@ class NavigatorState {
   NavigatorState({
     required this.context,
     required this.widget,
-  }) : scheduler = FrameworkRegistry.instance.getTaskScheduler(context);
+  }) : scheduler = ServicesRegistry.instance.getTaskScheduler(context);
 
   /*
   |--------------------------------------------------------------------------
@@ -504,7 +505,7 @@ class NavigatorState {
       var lastOpened = _historyStack.last;
 
       if (lastOpened.name == name) {
-        if (Utils.isKeyValueMapEqual(lastOpened.values, values)) {
+        if (fnIsKeyValueMapEqual(lastOpened.values, values)) {
           return;
         }
       }
